security/intel/stm: Make sure stm_resource_heap is consistent
When a parallel SMM relocation is being done, there is a good chance
that the value for stm_resource_heap is not consistent across processors.
Rather than holding (via a lock) processors until this value is set and
then flushing the cache so that all processors see the same value, this
solution moves the code such that all processors set it, thus maintaining
parallelism and keeping the code simple.

Change-Id: I2e5385c47124adcd99803337167984b6307af860
Signed-off-by: Eugene Myers <edmyers@tycho.nsa.gov>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/55629
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>